+++ python/paddle/fluid/tests/unittests/ipu/test_warpctc_op_ipu.py
@@ -98,11 +98,12 @@ class TestBase(IPUOpTest):
         label_length = paddle.static.data(
             name=self.feed_list[3], shape=self.feed_shape[3], dtype='int64'
         )
-        out = paddle.fluid.layers.warpctc(
+        out = paddle.nn.functional.ctc_loss(
             logits,
             labels,
             input_length=input_length,
             label_length=label_length,
+            reduction='mean',
             **self.attrs
         )
         loss = paddle.mean(out)
