+++ python/paddle/fluid/tests/unittests/test_dist_transpiler.py
@@ -1351,7 +1351,7 @@ class TestRemoteNce(TestDistLookupTableBase):
             )
         )
 
-        cost = fluid.layers.nce(
+        cost = paddle.static.nn.nce(
             input=input,
             label=label,
             num_total_classes=num_total_classes,
